move comp and helper
