+++ src/components/Selector/NativeSelect.tsx
@@ -2,7 +2,7 @@ import { Listbox } from "@headlessui/react";
 import { DrawerIcon } from "components/Icon";
 import { type AriaAttributes, forwardRef } from "react";
 import type { OptionType, ValKey } from "types/components";
-import { unpack } from "../form/helpers";
+import { unpack } from "helpers";
 import FocusableInput from "./FocusableInput";
 import { styles } from "./constants";
 import type { Classes } from "./types";
