+++ src/components/Selector/Selector.tsx
@@ -8,7 +8,7 @@ import {
   useController,
 } from "react-hook-form";
 import type { OptionType, ValKey } from "types/components";
-import { unpack } from "../form/helpers";
+import { unpack } from "helpers";
 import FocusableInput from "./FocusableInput";
 import { styles, valueKey } from "./constants";
 import type { Props } from "./types";
