+++ src/components/form/Radio.tsx
@@ -5,7 +5,7 @@ import {
   type PathValue,
   useFormContext,
 } from "react-hook-form";
-import { unpack } from "./helpers";
+import { unpack } from "helpers";
 import type { Classes } from "./types";
 
 type Props<T extends FieldValues, K extends Path<T>> = PropsWithChildren<{
