--- src/pages/Widget/Configurer/DonateMethods/index.ts
@@ -1,1 +1,0 @@
-export { DonateMethods } from "./DonateMethods";
