+++ src/pages/Widget/Configurer/schema.ts
@@ -1,7 +1,7 @@
 import type { SchemaShape } from "schemas/types";
-import type { Method } from "types/widget";
 import { type ObjectSchema, array, number, object } from "yup";
 import type { FormValues } from "./types";
+import { TDonateMethod } from "types/components";
 
 export const schema = object<any, SchemaShape<FormValues>>({
   endowment: object<any, SchemaShape<FormValues["endowment"]>>({
@@ -11,7 +11,7 @@ export const schema = object<any, SchemaShape<FormValues>>({
     "",
     "at least one payment option should be active",
     (values) => {
-      return values?.some((v) => !(v as Method).disabled);
+      return values?.some((v) => !(v as TDonateMethod).disabled);
     }
   ),
 }) as ObjectSchema<FormValues>;
